[refactor/#22] Refactor package structure
