--- src/main/java/com/api/jaebichuri/auction/dto/ProductDto.java
@@ -1,20 +1,0 @@
-package com.api.jaebichuri.auction.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ProductDto {
-
-    private Long productId;
-
-    private String productName;
-
-    private String productDescription;
-
-    private String startTime;
-
-    private String endTime;
-
-}
