--- src/main/java/com/api/jaebichuri/auction/entity/EndedAuctionImage.java
@@ -1,25 +1,0 @@
-package com.api.jaebichuri.auction.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class EndedAuctionImage {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String imageUrl;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ended_auction_id", nullable = false)
-    private EndedAuction endedAuction;
-
-}
