+++ src/main/java/com/api/jaebichuri/product/entity/AuctionProductImage.java
@@ -1,0 +1,24 @@
+package com.api.jaebichuri.auction.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class AuctionProductImage {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String imageUrl;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "auction_product_id", nullable = false)
+    private AuctionProduct auctionProduct;
+
+}
