+++ src/main/java/com/api/jaebichuri/screening/dto/ScreeningListDto.java
@@ -1,0 +1,17 @@
+package com.api.jaebichuri.auction.dto;
+
+import com.api.jaebichuri.auction.enums.AuctionCategory;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ScreeningListDto {
+
+    private Long screeningId;
+
+    private String productName;
+
+    private AuctionCategory auctionCategory;
+
+}
