+++ src/main/java/com/api/jaebichuri/screening/entity/AuctionScreening.java
@@ -1,0 +1,51 @@
+package com.api.jaebichuri.auction.entity;
+
+import com.api.jaebichuri.auction.enums.AuctionCategory;
+import com.api.jaebichuri.auction.enums.AuctionScreeningStatus;
+import com.api.jaebichuri.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@EntityListeners(AuditingEntityListener.class)
+public class AuctionScreening {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String productName;
+
+    @Column(columnDefinition = "TEXT", nullable = false)
+    private String productDescription;
+
+    @Column(nullable = false)
+    private Long desiredStartPrice;
+
+    @CreatedDate
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+    @Enumerated(EnumType.STRING)
+    private AuctionScreeningStatus screeningStatus;
+
+    @Enumerated(EnumType.STRING)
+    private AuctionCategory auctionCategory;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member seller;
+
+    public void setSeller(Member seller) {
+        this.seller = seller;
+    }
+}
