+++ src/main/java/com/api/jaebichuri/screening/entity/AuctionScreeningImage.java
@@ -1,0 +1,24 @@
+package com.api.jaebichuri.auction.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class AuctionScreeningImage {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String imageUrl;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "auction_screening_id", nullable = false)
+    private AuctionScreening auctionScreening;
+
+}
