+++ src/main/java/com/api/jaebichuri/screening/enums/AuctionScreeningStatus.java
@@ -1,0 +1,14 @@
+package com.api.jaebichuri.auction.enums;
+
+public enum AuctionScreeningStatus {
+
+    // 심사 대기
+    PENDING,
+
+    // 승인
+    APPROVED,
+
+    // 거절
+    REJECTED
+
+}
